blog/requirements.txt
black==23.12.0
Django==4.2.5
django-behaviors==0.5.1
djangorestframework==3.14.0
djlint==1.34.0
djoser==2.2.2
drf-spectacular==0.27.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-commas==2.1.0
flake8-return==1.2.0
gunicorn==21.2.0
isort==5.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
